chore: move heading components to /UI/headings
