+++ src/components/UI/Layout.tsx
@@ -1,20 +1,9 @@
-import { Container } from '@chakra-ui/react';
+import { Container, ContainerProps } from '@chakra-ui/react';
 import { FC } from 'react';
 
-interface Props {
-  bgGradient?: string;
-  h?: string;
-}
-
-export const Layout: FC<Props> = ({ bgGradient, children, h }) => {
+export const Layout: FC<ContainerProps> = ({ children, ...props }) => {
   return (
-    <Container
-      bgGradient={bgGradient}
-      h={h}
-      maxW={{ base: 'container.mobile', md: '600px' }}
-      px={5}
-      py={3}
-    >
+    <Container maxW={{ base: 'container.mobile', md: '600px' }} px={5} py={3} {...props}>
       {children}
     </Container>
   );
