+++ src/components/UI/PageText.tsx
@@ -1,9 +1,15 @@
-import { Text } from '@chakra-ui/react';
+import { Text, TextProps } from '@chakra-ui/react';
 import { FC } from 'react';
 
-export const PageText: FC = ({ children }) => {
+export const PageText: FC<TextProps> = ({ children, ...props }) => {
   return (
-    <Text color='brand.paragraph' fontSize='paragraph' fontWeight={300} lineHeight='24px'>
+    <Text
+      color='brand.paragraph'
+      fontSize='paragraph'
+      fontWeight={300}
+      lineHeight='24px'
+      {...props}
+    >
       {children}
     </Text>
   );
