+++ src/components/UI/headings/PageHeading.tsx
@@ -1,7 +1,7 @@
-import { Heading } from '@chakra-ui/react';
+import { Heading, HeadingProps } from '@chakra-ui/react';
 import { FC } from 'react';
 
-export const PageHeading: FC = ({ children }) => {
+export const PageHeading: FC<HeadingProps> = ({ children }) => {
   return (
     <Heading as='h1' color='brand.heading' fontSize='h1' fontWeight={200} lineHeight='48px' mb={4}>
       {children}
