+++ src/components/UI/headings/index.ts
@@ -1,0 +1,2 @@
+export { PageHeading } from './PageHeading';
+export { PageSubheading } from './PageSubheading';
