+++ src/components/UI/index.ts
@@ -1,5 +1,5 @@
+export { PageHeading } from './headings/PageHeading';
 export { Layout } from './Layout';
 export { NavMobile } from './NavMobile';
-export { PageHeading } from './PageHeading';
 export { PageText } from './PageText';
 export { PlaceholderImage } from './PlaceholderImage';
